Moved the SharedCollectionButton into the Collection Options file
