+++ Jewel/Views/Options/CollectionOptions.swift
@@ -112,3 +112,32 @@ struct CollectionOptions: View {
   }
 }
 
+struct ShareCollectionButton: View {
+  
+  @EnvironmentObject var app: AppEnvironment
+  
+  var collection: Collection
+  
+  @State private var showSharing: Bool = false
+  
+  private var isOnRotation: Bool {
+    self.collection.id == self.app.state.library.onRotation.id
+  }
+  
+  var body: some View {
+    Button(action: {
+      self.showSharing = true
+    }) {
+      HStack {
+        Image(systemName: "square.and.arrow.up")
+          .frame(width: Constants.optionsButtonIconWidth)
+        Text("Share \(self.isOnRotation ? "On Rotation" : "Collection")")
+      }
+    }
+    .sheet(isPresented: self.$showSharing) {
+      ShareSheetLoader(collection: self.collection)
+        .environmentObject(self.app)
+    }
+  }
+}
+
